requirements.txt
python-decouple
openai
asyncio
fastapi
uvicorn 